Add home page before settings
